pyproject.toml
[tool.poetry]
name = "payment-system"
version = "0.1.0"
description = ""
authors = ["ElenaGlu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.6"
uvloop = "^0.20.0"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"
sqlalchemy = {version = "^2.0.32", extras = ["asyncio"]}
python-multipart = "^0.0.9"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.0"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.3.2"
yarl = "^1.9.4"
mypy = "^1.11.1"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
